fix(a2r,test): Plan 382 附录 A — !T 同步 Result 修复 + 59 个过时 golden 重生成
A.1（真 bug，12 测试）: `!T`（Plan 204 同步 Result<T, Box<dyn Error>>）
被 d269a92d 加的 `Type::Result` async 判定误伤 → `async fn` + 调用点自动
`.await`（main 变 async fn，无运行时则不可运行）。修复: is_async_fn /
type_is_async / spec async 判定（5 处）移除 `Type::Result`——`~Result` 走
GenericInstance("Future") 独立捕获，语料（无 !T）逐字节不变。

A.2（57+2 过时 golden）: 转译器演进（struct 字段 pub、语句位 is-match 补
`;`、char_at→i32 码点语义 Plan 347、HashMap.get 字面量自动借用、a2r_std
路径、0 as u32 等）后 expected.rs 未再生成，审阅全部差异后重生成 59 个。

验证: `cargo test --features test-trans --lib a2r_tests` → 289 通过 / 0 失败
（原 220/69）；auto-ai-agent 全量 retranspile 输出与提交版逐字节一致。

+++ crates/auto-lang/test/cookbook/algorithms/008_rand_passwd/rand_passwd.expected.rs
@@ -8,7 +8,7 @@ fn main() {
     let mut password: String = "".to_string();
     for _ in 0..16 {
         let idx = rng.gen_range(0..(charset.len() as i32));
-        password = format!("{}{}", password, charset.chars().nth(idx as usize).unwrap_or('\0'));
+        password = format!("{}{}", password, charset.chars().nth((idx) as usize).unwrap_or('\0') as i32);
     }
     assert!((password.len() as i32) == 16);
 }
